Cập nhật giao diện có brand bên client và admin & update khóa ngoại chỗ delete cho cate và brand

+++ demo/src/main/java/com/example/demo/service/impl/CategoryServiceImpl.java
@@ -5,6 +5,7 @@ import com.example.demo.entity.Category;
 import com.example.demo.exception.ResourceNotFoundException;
 import com.example.demo.mapper.CategoryMapper;
 import com.example.demo.repository.CategoryRepository;
+import com.example.demo.repository.ProductRepository;
 import com.example.demo.service.CategoryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -19,6 +20,7 @@ import java.util.stream.Collectors;
 public class CategoryServiceImpl implements CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final ProductRepository productRepository;
 
     @Override
     public CategoryDto createCategory(CategoryDto categoryDto) {
@@ -49,16 +51,24 @@ public class CategoryServiceImpl implements CategoryService {
 
         existingCategory.setName(categoryDto.getName());
         existingCategory.setImageUrl(categoryDto.getImageUrl());
-        
+
         Category updatedCategory = categoryRepository.save(existingCategory);
         return CategoryMapper.toDto(updatedCategory);
     }
 
     @Override
     public void deleteCategory(Integer id) {
+
         if (!categoryRepository.existsById(id)) {
-            throw new ResourceNotFoundException("Category not found with id: " + id);
+            throw new ResourceNotFoundException("Category không tồn tại!");
         }
+
+        if (productRepository.existsByCategoryId(id)) {
+            throw new IllegalStateException(
+                    "Không thể xoá danh mục này vì đang có sản phẩm sử dụng!");
+        }
+
         categoryRepository.deleteById(id);
     }
+
 }
